requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.1.0
Automat==20.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
censys==2.1.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
commonmark==0.9.1
constantly==15.1.0
cryptography==39.0.0
dnspython==2.3.0
exceptiongroup==1.1.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
outcome==1.2.0
packaging==23.0
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
rfc3986==1.5.0
rich==13.1.0
selenium==4.3.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.4.0
typing_extensions==4.3.0
urllib3==1.26.12
urllib3-secure-extra==0.1.0
wsproto==1.2.0
zope.interface==5.4.0
